+++ e_out_of_date/command.py
@@ -8,7 +8,7 @@ import sys
 
 from e_out_of_date import (
     Analyzer, InstalledPackageVersions, AvailablePackageVersions, ReportedUpdateTypes,
-    PackageVersionClassifications
+    PackageVersionClassifications, ignore_pkg_resources_warnings
 )
 
 
@@ -28,6 +28,8 @@ def main(args):
                         action='store_true')
     parser.add_argument('freeze_output', nargs='?')
     args = parser.parse_args(args)
+
+    ignore_pkg_resources_warnings()
 
     try:
         types = ReportedUpdateTypes(types=args.types)