+++ e_out_of_date/util.py
@@ -1,0 +1,7 @@
+import warnings
+
+from pkg_resources import PkgResourcesDeprecationWarning
+
+
+def ignore_pkg_resources_warnings():
+    warnings.filterwarnings("ignore", category=PkgResourcesDeprecationWarning)